Keep it separate and internal for now
