+++ packages/liveblocks-react-ui/src/primitives/AiMessage/index.tsx
@@ -1,28 +1,15 @@
-import type {
-  AiChatMessage,
-  AiToolInvocationPart,
-  AiToolInvocationProps,
-  JsonObject,
-  ToolResultData,
-  ToolResultResponse,
-} from "@liveblocks/core";
-import { kInternal } from "@liveblocks/core";
-import { useClient } from "@liveblocks/react";
-import { useSignal } from "@liveblocks/react/_private";
 import { Slot } from "@radix-ui/react-slot";
-import type { FunctionComponent } from "react";
-import { forwardRef, useCallback, useMemo } from "react";
+import { forwardRef, useMemo } from "react";
 
 import { ErrorBoundary } from "../../utils/ErrorBoundary";
 import { Markdown } from "../Markdown";
-import { AiToolInvocationContext } from "./contexts";
+import { AiMessageToolInvocation } from "./tool-invocation";
 import type {
   AiMessageContentComponents,
   AiMessageContentProps,
 } from "./types";
 
 const AI_MESSAGE_CONTENT_NAME = "AiMessageContent";
-const AI_MESSAGE_TOOL_INVOCATION_NAME = "AiMessageToolInvocation";
 
 const defaultMessageContentComponents: AiMessageContentComponents = {
   TextPart: ({ part }) => {
@@ -39,83 +26,6 @@ const defaultMessageContentComponents: AiMessageContentComponents = {
     );
   },
 };
-
-/* -------------------------------------------------------------------------------------------------
- * ToolInvocationPart
- * -----------------------------------------------------------------------------------------------*/
-
-function StableRenderFn(props: {
-  renderFn: FunctionComponent<
-    AiToolInvocationProps<JsonObject, ToolResultData>
-  >;
-  props: AiToolInvocationProps<JsonObject, ToolResultData>;
-}) {
-  return props.renderFn(props.props);
-}
-
-/**
- * @internal
- */
-function AiMessageToolInvocation({
-  message,
-  part,
-}: {
-  message: AiChatMessage;
-  part: AiToolInvocationPart;
-}) {
-  const client = useClient();
-  const ai = client[kInternal].ai;
-  const tool = useSignal(ai.signals.getToolΣ(part.name, message.chatId));
-
-  const respond = useCallback(
-    (result: ToolResultResponse<ToolResultData>) => {
-      if (part.stage === "receiving") {
-        console.log(
-          `Ignoring respond(): tool '${part.name}' (${part.invocationId}) is still receiving`
-        );
-      } else if (part.stage === "executed") {
-        console.log(
-          `Ignoring respond(): tool '${part.name}' (${part.invocationId}) has already executed`
-        );
-      } else {
-        ai.setToolResult(
-          message.chatId,
-          message.id,
-          part.invocationId,
-          result.data
-          // TODO Pass in AiGenerationOptions here?
-        );
-      }
-    },
-    [ai, message.chatId, message.id, part.stage, part.name, part.invocationId]
-  );
-
-  const props = useMemo(() => {
-    const { type: _, ...rest } = part;
-    return {
-      ...rest,
-      respond,
-      types: undefined as never,
-      [kInternal]: {
-        execute: tool?.execute,
-      },
-    };
-  }, [part, respond, tool?.execute]);
-
-  if (tool?.render === undefined) return null;
-  return (
-    <AiToolInvocationContext.Provider value={props}>
-      <StableRenderFn
-        renderFn={
-          tool.render as FunctionComponent<
-            AiToolInvocationProps<JsonObject, ToolResultData>
-          >
-        }
-        props={props}
-      />
-    </AiToolInvocationContext.Provider>
-  );
-}
 
 /**
  * --------------------------------------------------------------------------
@@ -173,11 +83,7 @@ const AiMessageContent = forwardRef<HTMLDivElement, AiMessageContentProps>(
 
 if (process.env.NODE_ENV !== "production") {
   AiMessageContent.displayName = AI_MESSAGE_CONTENT_NAME;
-  AiMessageToolInvocation.displayName = AI_MESSAGE_TOOL_INVOCATION_NAME;
 }
 
 // NOTE: Every export from this file will be available publicly as AiMessage.*
-export {
-  AiMessageContent as Content,
-  AiMessageToolInvocation as ToolInvocation,
-};
+export { AiMessageContent as Content };
